Remove duplicate property storage
We represent tree nodes by two different objects, a "PhyloNode" and a "Node", although these names are not formalised. The former holds information necessary to render the phylogenetic tree and is solely used by `PhyloTree` while the latter holds more general information. They are linked: `phyloNode = node.n` and `node = phyloNode.shell`.

This commit tackles deeply embedded technical debt where each object held a copy of the same data. Specifically,
* `children` is now only stored on a Node
* `parent` is similarly stored on a Node
* `terminal` is removed from PhyloNode (we can refer to node.hasChildren)

There is (at least) one remaining duplicated property - `inView` - which is set on <Node> when the dataset loads due to the corresponding <PhyloNode> not yet existing. This is left for future work as the concept of `inView` may change as we develop further methods for tree zoom / pan.