WIP Exploded trees
Introduces the ability to cut trees at various branches and display the resulting trees as subtrees. These cut points are here defined to be a change in a (user-selected) state trait, which are simply the non-continuous colourings (to start with).

Briefly, given a parent node (A) with children (B,C), and a change in trait moving from (A->B), we add B to the children of the overall tree root, and update the children of A to be only C. We store the original children (B,C) so that we can return to the starting state.

This implementation, specifically the ability for PhyloTree to update the node ordering, should allow for branch rotations in the future.

There are a number of future improvements to be made in subsequent commits:
* The subtrees should be ordered on trait (or we may wish to make this a toggle)
* the stem length of subtrees is inconsistent
* exploded trait cannot be stored in URL state
* we should limit the explode options to those colourings which are defined on internal nodes (currently a warning is shown)
* genotype should be available to explode on
* remove console logging introduced here
* consider undefined (internal) traits, e.g. currently nodes (X -> undefined -> Y) would not be exploded
* multiple trees are not considered here

Prior art: I was first made aware of this style of tree rendering by Gytis Dudas. An implementation was attempted by @frogsquire<https://github.com/frogsquire> in https://github.com/nextstrain/auspice/pull/1105 and this commit uses ideas introduced there.