Introduce descriptions of major/minor parents
Given a subtree root node, whether formed from multiple trees in the dataset or by exploding a tree, we wish to include information about the parent node. This may include where in another tree this (sub-)tree originated (referred to here as “major”) or a recombination donor (“minor”). This differs from the parent used to describe the overall tree structure, which is an implementation detail.

We use the major parent (if defined) to define the branch length of the subtree root, which is otherwise ambiguous. In the future we may also render a connecting line to indicate this relationship. The current commit does not consider “minor” parents, but one possible display of this can be seen at https://cse512-21s.github.io/FP-Pathogen-Phylogenies/ (Lewinsohn, Paredes, Russell and Wagner).

This commit only implements this concept for exploded trees, however the intention is for this to be user definable via a `branch_attrs.parent` annotation. The following situations are possible:

(i) Single tree, not exploded: Minor parents (via  `branch_attrs.parent`) may be used to visualise recombination events (see above). Major parents are not applicable.
(ii) Multi tree, not exploded: Major parents (via  `branch_attrs.parent`) can indicate the placement of the subtree (in another tree). Minor parents may also be applicable here.
(iii) Single tree, exploded:  `branch_attrs.parent` is ignored. We set the major parent of each subtree to be the branch it originated from.
(iv) Multi tree, exploded: Same as (iii).