Add Multiplot panel
Add the skeleton multiplot panel that includes the legend and a
container for the SVG. Currently acts as a third "big" panel that can
span the full page or be turned into a half panel for grid display.
Legend is hidden when the panel is in the grid display with the tree.