Fetch and load multiplot sidecar JSON file
The multiplot sidecar JSON file should have a "_multiplot" suffix in the
filename. The JSON is expected to have a "collections" array that
contains individual collections with their respective display configs
and measurements data. See more details about JSON schema within the
augur repo.

On loading the data, an id and jitter value are added to each
measurement. The id is just the measurement's array index but is added
to ensure a stable id even if the list of measurements has been filtered.
The id provides a way to create a stable DOM id that can be used to
select measurements individually. The jitter will be used as the y-value
for each dot so that the displayed measurements are spread apart
vertically within each subplot.

The collections array and default collection object are added to the
multiplot redux store. The display config for the default collection
object is parsed and added to the controls redux store. Ordering of
collections and their groupings are preserved by storing their
key/title pairs in Map objects.