multiplot: Add D3 SVG
Uses the display config and data from the redux store to create an
SVG using D3 in the multiplot panel.

Only displays measurements that are linked to samples currently visible
on the tree, so each measurement's "strain" value must exactly match
the strain name of the sample in the tree. This allows the multiplot to
share the "Color By" of the tree and ensures that filters of the tree
also filters data of the multiplot.