multiplot: add filter
Adds a searchable filter dropdown in the sidebar control panel to filter
the multiplot data by fields and values. Currently excludes the fields
"multiplotJitter" and "multiplotId" because they are only used as helper
fields for the visualization. Also excludes the "value" field since it
can potentially bloat the dropdown menu with numeric values that I do
not foresee as a useful filter.

Uses the existing FilterBadge component to display the summary of field
filters in the sidebar and individual value filters in the top section
of the multiplot panel.

Future work may include displaying the counts of measurements shown and
the counts of measurements included in each value filter.